refactor: changed static back to dist
