+++ vite.config.ts
@@ -3,8 +3,5 @@ import react from "@vitejs/plugin-react";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  build: {
-    outDir: "static",
-  },
   plugins: [react()],
 });
